fullstack-movie-search/Multimodal-Movie-Script-Search-Engine/backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
torch==2.0.1
transformers==4.33.2
sentence-transformers==2.2.2
pillow==10.0.0
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0
